use schema defined by cr8 tool for benchmark api
