+++ .gitignore
@@ -31,3 +31,11 @@ hs_err_pid*
 out/
 venv/
 py/notebooks/.ipynb_checkpoints/
+
+# Python stuff
+bin/
+parts/
+__pycache__/
+*.egg-info/
+develop-eggs/
+.installed.cfg
