+++ api/buildout.cfg
@@ -8,4 +8,3 @@ parts = app
 [app]
 recipe = zc.recipe.egg:script
 eggs = benchmark-api
-       crate
