+++ api/versions.cfg
@@ -7,9 +7,9 @@ Werkzeug = 0.11.3
 aniso8601 = 1.1.0
 itsdangerous = 0.24
 six = 1.10.0
+ourllib3 = 1.16
+zc.buildout = 2.5.2
 zc.recipe.egg = 2.0.3
-setuptools = 19.4
-zc.buildout = 2.5.0
 
 # Required by:
 # Jinja2==2.8
@@ -22,3 +22,11 @@ python-dateutil = 2.4.2
 # Required by:
 # Flask-RESTful==0.3.5
 pytz = 2015.7
+
+# Required by:
+# benchmark-api==0.1.0
+crate = 0.16.3
+
+# Required by:
+# crate==0.16.3
+setuptools = 26.0.0
